Extract withCleanParamNames ext method
