+++ compiler/src/dotty/tools/dotc/core/Decorators.scala
@@ -5,7 +5,7 @@ package core
 import scala.annotation.tailrec
 import scala.collection.mutable.ListBuffer
 
-import Contexts.*, Names.*, Phases.*, Symbols.*
+import Contexts.*, Names.*, Phases.*, Symbols.*, Types.*
 import printing.{ Printer, Showable }, printing.Formatting.*, printing.Texts.*
 import transform.MegaPhase
 import reporting.{Message, NoExplanation}
@@ -79,6 +79,24 @@ object Decorators {
       }
       NoSymbol
     }
+
+  extension (tp: Type)
+    /** Replace synthetic parameter names (`x$0`, `x$1`, ...) of any method
+     *  type group in `tp` (recursing through curried `MethodType`s and
+     *  through `PolyType` result types) with dollar-free names (`x0`,
+     *  `x1`, ...). Lets a printed signature be reused as a valid Scala
+     *  identifier - e.g. in stub implementations or "method is not
+     *  defined" diagnostics.
+     */
+    def withCleanParamNames(using Context): Type = tp match
+      case mt: MethodType if mt.allParamNamesSynthetic =>
+        val newNames = mt.paramNames.zipWithIndex.map((_, i) => termName("x" + i))
+        mt.derivedLambdaType(newNames, mt.paramInfos, mt.resType.withCleanParamNames)
+      case mt: MethodType =>
+        mt.derivedLambdaType(mt.paramNames, mt.paramInfos, mt.resType.withCleanParamNames)
+      case pt: PolyType =>
+        pt.derivedLambdaType(pt.paramNames, pt.paramInfos, pt.resType.withCleanParamNames)
+      case _ => tp
 
   inline val MaxFilterRecursions = 10
